Move omega_sigma from scil_evaluate_connectivity... to stats.matrix_stats
