+++ scilpy/stats/tests/test_matrix_stats.py
@@ -5,3 +5,6 @@ def test_ttest_two_matrices():
     # toDo
     pass
 
+
+def test_omega_sigma():
+    pass
